chore: renaming index files
