+++ site/src/components/CliAuthToken/CliAuthToken.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from "@testing-library/react"
 import React from "react"
-import { render } from "../../testHelpers"
+import { render } from "../../testHelpers/renderHelpers"
 import { CliAuthToken } from "./CliAuthToken"
 
 describe("CliAuthToken", () => {
