+++ site/src/components/EmptyState/EmptyState.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from "@testing-library/react"
 import React from "react"
-import { render } from "../../testHelpers"
+import { render } from "../../testHelpers/renderHelpers"
 import { EmptyState } from "./EmptyState"
 
 describe("EmptyState", () => {
