+++ site/src/components/ResetPasswordDialog/ResetPasswordDialog.stories.tsx
@@ -1,6 +1,6 @@
 import { Story } from "@storybook/react"
 import React from "react"
-import { MockUser } from "../../testHelpers"
+import { MockUser } from "../../testHelpers/renderHelpers"
 import { ResetPasswordDialog, ResetPasswordDialogProps } from "./ResetPasswordDialog"
 
 export default {
