+++ site/src/components/UserCell/UserCell.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, Story } from "@storybook/react"
 import React from "react"
-import { MockUser, MockUserAgent } from "../../testHelpers"
+import { MockUser, MockUserAgent } from "../../testHelpers/renderHelpers"
 import { UserCell, UserCellProps } from "./UserCell"
 
 export default {
