+++ site/src/components/UserCell/UserCell.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from "@testing-library/react"
 import React from "react"
-import { MockUser, MockUserAgent, WrapperComponent } from "../../testHelpers"
+import { MockUser, MockUserAgent, WrapperComponent } from "../../testHelpers/renderHelpers"
 import { UserCell, UserCellProps } from "./UserCell"
 
 namespace Helpers {
