+++ site/src/forms/CreateWorkspaceForm.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react"
 import React from "react"
-import { MockOrganization, MockTemplate, MockWorkspace } from "../testHelpers"
+import { MockOrganization, MockTemplate, MockWorkspace } from "../testHelpers/renderHelpers"
 import { CreateWorkspaceForm } from "./CreateWorkspaceForm"
 
 describe("CreateWorkspaceForm", () => {
