+++ site/src/pages/TemplatesPages/OrganizationPage/TemplatePage/CreateWorkspacePage.tsx
@@ -3,7 +3,7 @@ import { useSelector } from "@xstate/react"
 import React, { useCallback, useContext } from "react"
 import { useNavigate, useParams } from "react-router-dom"
 import useSWR from "swr"
-import * as API from "../../../../api"
+import * as API from "../../../../api/api"
 import * as Types from "../../../../api/types"
 import { ErrorSummary } from "../../../../components/ErrorSummary/ErrorSummary"
 import { FullScreenLoader } from "../../../../components/Loader/FullScreenLoader"
