+++ site/src/xServices/auth/authXService.ts
@@ -1,5 +1,5 @@
 import { assign, createMachine } from "xstate"
-import * as API from "../../api"
+import * as API from "../../api/api"
 import * as Types from "../../api/types"
 import * as TypesGen from "../../api/typesGenerated"
 import { displaySuccess } from "../../components/GlobalSnackbar/utils"
