+++ site/src/xServices/buildInfo/buildInfoXService.ts
@@ -1,5 +1,5 @@
 import { assign, createMachine } from "xstate"
-import * as API from "../../api"
+import * as API from "../../api/api"
 import * as Types from "../../api/types"
 
 export interface BuildInfoContext {
