+++ site/src/xServices/terminal/terminalXService.ts
@@ -1,5 +1,5 @@
 import { assign, createMachine } from "xstate"
-import * as API from "../../api"
+import * as API from "../../api/api"
 import * as Types from "../../api/types"
 
 export interface TerminalContext {
