+++ site/src/xServices/users/usersXService.ts
@@ -1,5 +1,5 @@
 import { assign, createMachine } from "xstate"
-import * as API from "../../api"
+import * as API from "../../api/api"
 import { ApiError, FieldErrors, isApiError, mapApiErrorToFieldErrors } from "../../api/errors"
 import * as Types from "../../api/types"
 import * as TypesGen from "../../api/typesGenerated"
